refactor: split constant propagation statements
